Fix eslint error
